fix infinite looping
when merging the child-context into the requester's context after fetching relationships,
the requester's context gets the newly fetched relationships and tries to fault-in the
relationship we just created. This triggers another fetch on the server and it starts from the beginning.
By storing that we're currently saving, we bypass the server update.

I'm not exactly sure why this happens for relationships and not for the values.